# openapi.yaml
openapi: 3.1.0
info:
  title: fitbase API
  description: |
    Self-hosted cycling fitness data API. Designed for LLM agent consumption.
    All responses use a consistent `{"data": ..., "error": ...}` envelope.
    Field names always include units (e.g., `duration_secs`, `power_watts`).

    Many sync endpoints support `?stream=1` which switches the response to
    Server-Sent Events (`text/event-stream`) for real-time progress updates.
    SSE event types: `start` (total/pending counts), `file` (per-file progress),
    `done` (final counts), `error` (on failure).
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/workouts/{id}:
    get:
      summary: Get workout
      parameters:
        - $ref: "#/components/parameters/WorkoutID"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete workout
      parameters:
        - $ref: "#/components/parameters/WorkoutID"
      responses:
        "204":
          description: Workout deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    WorkoutID:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Workout ID (first 16 hex chars of SHA-256 of FIT file content)
...
  schemas:
    Workout:
      type: object
      properties:
        id:                     { type: string, example: "a3f9c2d1e4b5f6a7" }
        filename:               { type: string }
        recorded_at:            { type: string, format: date-time }
        sport:                  { type: string, enum: [cycling, running, swimming, hiking, other] }
        duration_secs:          { type: integer }
        elapsed_secs:           { type: integer }
        distance_meters:        { type: number }
        elevation_gain_meters:  { type: number }
        avg_power_watts:        { type: number, nullable: true }
        max_power_watts:        { type: number, nullable: true }
        normalized_power_watts: { type: number, nullable: true }
        avg_heart_rate_bpm:     { type: integer, nullable: true }
        max_heart_rate_bpm:     { type: integer, nullable: true }
        avg_cadence_rpm:        { type: integer, nullable: true }
        avg_speed_mps:          { type: number }
        tss:                    { type: number, nullable: true }
        intensity_factor:       { type: number, nullable: true }
        is_indoor:              { type: boolean }
        route_id:               { type: string, nullable: true }
        created_at:             { type: string, format: date-time }
...
    WorkoutResponse:
      type: object
      properties:
        data:  { $ref: "#/components/schemas/Workout" }
        error: { type: string, nullable: true }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...